protocols

Module rep3_ring

REP3 Ring

This module implements the rep3 share and combine operations for rings

Modules

arithmetic

Arithmetic

binary

Binary

conversion

Conversions

detail

Implementation of the low-depth binary addition This module provides functions for performing low-depth binary addition

Structs

Rep3RingShare

This type represents a replicated shared value. Since a replicated share of a ring element contains additive shares of two parties, this type contains two ring elements.

Functions

combine_ring_element

Reconstructs a ring element from its arithmetic replicated shares.

combine_ring_element_binary

Reconstructs a ring element from its binary replicated shares.

combine_ring_elements

Reconstructs a vector of ring elements from its arithmetic replicated shares.

combine_ring_elements_binary

Reconstructs a vector of ring elements from its binary replicated shares.

share_ring_element

The internal state of the REP3 protocol Secret shares a ring element using replicated secret sharing and the provided random number generator. The ring element is split into three additive shares, where each party holds two. The outputs are of type Rep3RingShare.

share_ring_element_binary

Secret shares a ring element using replicated secret sharing and the provided random number generator. The ring element is split into three binary shares, where each party holds two. The outputs are of type Rep3RingShare.

share_ring_elements

Secret shares a vector of ring elements using replicated secret sharing and the provided random number generator. The ring elements are split into three additive shares each, where each party holds two. The outputs are of type Rep3RingShare.

share_ring_elements_binary

Secret shares a vector of ring elements using replicated secret sharing and the provided random number generator. The ring elements are split into three binary shares each, where each party holds two. The outputs are of type Rep3RingShare.

Type Aliases

Rep3BitShare

Shorthand type for a secret shared bit.

SeedRng

The Rng used for expanding compressed Shares