protocols::rep3_ring

Module arithmetic

Arithmetic

This module contains operations with arithmetic shares

Functions

add

Performs addition between two shared values.

add_assign

Performs addition between two shared values in place

add_assign_public

Performs addition between a shared value and a public value in place.

add_public

Performs addition between a shared value and a public value.

add_vec

Performs addition between two shared values.

add_vec_assign

Performs element-wise addition of two vectors of shared values in place.

local_mul_vec

Performs element-wise multiplication of two vectors of shared values. DOES NOT PERFORM RESHARE

mul_assign_public

Performs multiplication of a shared value and a public value.

mul_public

Performs multiplication of a shared value and a public value.

neg

Negates a shared value.

open

Performs the opening of a shared value and returns the equivalent public value.

open_bit

Performs the opening of a shared value and returns the equivalent public value.

open_vec

Performs the opening of a shared value and returns the equivalent public value.

open_vec_bit

Qamboo: create

promote_to_trivial_share

Transforms a public value into a shared value: [a] = a.

reshare_vec

Performs a reshare on all shares in the vector.

sub

Performs subtraction between two shared values, returning a - b.

sub_assign

Performs subtraction between two shared values in place.

sub_public_by_shared

Performs subtraction between a shared value and a public value, returning public - shared.

sub_shared_by_public

Performs subtraction between a shared value and a public value, returning shared - public.

sub_vec_assign

Performs element-wise subtraction of two vectors of shared values in place.

Type Aliases

RingShare

Type alias for a Rep3RingShare which is used for both arithmetic and binary shares.