protocols::rep3_ring

Module binary

Binary

This module contains operations with binary shares

Functions

and

Performs a bitwise AND operation on two shared values.

and_vec

Performs element-wise bitwise AND operation on the provided shared values.

and_vec_bit

bit and

and_with_public

Performs a bitwise AND operation on a shared value and a public value.

and_with_public_many

Qamboo add : Performs a bitwise AND operation on a vector of shared values and a public value.

cmux

Computes a CMUX: If c is 1, returns x_t, otherwise returns x_f.

cmux_many

Computes a CMUX for multiple inputs.

is_zero

Computes a binary circuit to check whether the replicated binary-shared input x is zero or not. The output is a binary sharing of one bit.

open

Performs the opening of a shared value and returns the equivalent public value.

open_vec

Performs the opening of a vector of shared values and returns the equivalent public values.

or

Performs a bitwise OR operation on two shared values.

or_public

Performs a bitwise OR operation on a shared value and a public value.

or_tree

Compute a OR tree of the input vec

or_vec

Performs a bitwise OR operation on two vector of shared values.

promote_to_trivial_share

Transforms a public value into a shared value: [a] = a.

shift_l_public

Shifts a share by a public value F to the left.

shift_r_public

Shifts a share by a public value F to the right.

xor

Performs a bitwise XOR operation on two shared values.

xor_public

Performs a bitwise XOR operation on a shared value and a public value.