protocols::rep3_ring

Module conversion

Conversions

This module contains conversions between share types

Functions

a2b

Transforms the replicated shared value x from an arithmetic sharing to a binary sharing. I.e., x = x_1 + x_2 + x_3 gets transformed into x = x’_1 xor x’_2 xor x’_3.

a2b_many

Transforms the replicated shared value x from an arithmetic sharing to a binary sharing. I.e., x = x_1 + x_2 + x_3 gets transformed into x = x’_1 xor x’_2 xor x’_3.

b2a

Transforms the replicated shared value x from a binary sharing to an arithmetic sharing. I.e., x = x_1 xor x_2 xor x_3 gets transformed into x = x’_1 + x’_2 + x’_3.

b2a_many

A variant of b2a that operates on vectors of shared values instead.

bit_inject

Translates one shared bit into an arithmetic sharing of the same bit. I.e., the shared bit x = x_1 xor x_2 xor x_3 gets transformed into x = x’_1 + x’_2 + x’_3, with x being either 0 or 1.

bit_inject_many

Translates a vector of shared bits into a vector of arithmetic sharings of the same bits. See bit_inject for details.